se cambia el problema
se arregla el problema con  una cosa

+++ src/main/java/co/edu/ucentral/repository/IUsuariosRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import co.edu.ucentral.entidades.Usuario;
 
-public interface IUsuariosRepository extends JpaRepository<Usuario, Long>{
+public interface IUsuariosRepository extends JpaRepository<Usuario, Integer>{
 
 	Usuario findByNumeroIdentificacionAndPassword(Integer numeroIdentificacion, String pass);
 }
